+++ NEWS.md
@@ -15,6 +15,7 @@
 
 * To reduce interface complexity, `find_parameters()` and `get_parameters()` for objects of class `aovlist` now return the elements `$conditional` and `$random`, to be in line with other supported objects.
 * The argument `resp` in `get_response()` was renamed to `select`, to have a more clear verb (#114).
+* The functions `find_variables()` and `find_terms()` were flipped, because what previously was considered as "term" was actually a "variable", and vice versa.
 
 ## Changes to functions